requirements.txt
emoji==2.12.1
Faker==28.0.0
pika==1.3.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
PyYAML==6.0.2
streamlit==1.37.1
